New overview page implemented.
