+++ lessons/module/module-01/module-01.mdx
@@ -3,6 +3,7 @@ filterBy: lesson-1
 slug: module-01
 type: module
 next: '/tezos/storyline'
+title: Basics of SmartPy
 ---
 # This is the overview for module-1
 lorem ipsum lorem ipsumlorem ipsumlorem ipsumlorem ipsumlorem ipsum
