+++ lessons/module/module-02/module-02.mdx
@@ -3,5 +3,6 @@ filterBy: lesson-2
 slug: module-02
 type: module
 next: /lesson/module-02/chapter-01
+title: Inter-contract calling
 ---
 This is the overview for module-2
