+++ lessons/module/module-03/module-03.mdx
@@ -3,5 +3,6 @@ filterBy: lesson-3
 slug: module-03
 type: module
 next: /lesson/module-03/chapter-01
+title: FA2 Tokens
 ---
 This is the overview for module-3
