+++ lessons/module/module-04/module-04.mdx
@@ -3,5 +3,6 @@ filterBy: lesson-4
 slug: module-04
 type: module
 next: /lesson/module-04/chapter-01
+title: Taquito
 ---
 This is the overview for module-4
